- Added multiplatform CI (Windows/Linux/MacOS) - No tests on windows - dll problems - Fixed some compiler/linter warnings - Simplified CI scripts - Moved fetching & building gtest to cmake

+++ libecho/src/AudioOutput.cc
@@ -16,7 +16,7 @@ void AudioOutput::enqueueData(const char *data, int length) {
 }
 
 void AudioOutput::tryWriteData() {
-    size_t bytesToWrite = std::min(qStream->bytesFree(), buffer.size());
+    int bytesToWrite = std::min(qStream->bytesFree(), buffer.size());
     if (bytesToWrite == 0) {
         return;
     }
